+++ transaction-context/src/lib.rs
@@ -288,7 +288,7 @@ impl<'ix_data> TransactionContext<'ix_data> {
         instruction_accounts: Vec<InstructionAccount>,
         deduplication_map: Vec<u16>,
         instruction_data: Cow<'ix_data, [u8]>,
-        parent_index: Option<u32>,
+        parent_index: Option<u16>,
     ) -> Result<(), InstructionError> {
         debug_assert_eq!(deduplication_map.len(), MAX_ACCOUNTS_PER_TRANSACTION);
         let trace_len = self.instruction_trace.len();
@@ -299,13 +299,13 @@ impl<'ix_data> TransactionContext<'ix_data> {
             .last_mut()
             .ok_or(InstructionError::CallDepth)?;
 
-        instruction.program_account_index_in_tx = program_index as u64;
+        instruction.program_account_index_in_tx = program_index as u32;
         instruction.configure_vm_slices(
             instruction_index as u64,
             instruction_accounts.len(),
             instruction_data.len() as u64,
         );
-        instruction.index_of_parent_instruction = parent_index.unwrap_or(u32::MAX);
+        instruction.index_of_parent_instruction = parent_index.unwrap_or(u16::MAX);
         self.deduplication_maps
             .push(deduplication_map.into_boxed_slice());
         self.instruction_accounts
@@ -352,7 +352,7 @@ impl<'ix_data> TransactionContext<'ix_data> {
                 .instruction_trace
                 .last_mut()
                 .ok_or(InstructionError::CallDepth)?;
-            instruction.nesting_level = nesting_level as u32;
+            instruction.nesting_level = nesting_level as u16;
         }
         let index_in_trace = self.get_instruction_trace_length();
         if index_in_trace >= self.instruction_trace_capacity {
